Move metadata utility classes to the smae package for a clearer overview
