+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/AggregateUtils.java
@@ -2,7 +2,7 @@ package org.molgenis.data.elasticsearch;
 
 import org.molgenis.data.meta.model.Attribute;
 
-import static org.molgenis.data.support.EntityTypeUtils.isReferenceType;
+import static org.molgenis.data.util.EntityTypeUtils.isReferenceType;
 
 public class AggregateUtils
 {
