+++ molgenis-data-postgresql/src/main/java/org/molgenis/data/postgresql/PostgreSqlRepositoryCollection.java
@@ -5,7 +5,7 @@ import org.molgenis.data.*;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.support.AbstractRepositoryCollection;
-import org.molgenis.data.support.AttributeUtils;
+import org.molgenis.data.util.AttributeUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -26,13 +26,13 @@ import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
 import static org.molgenis.data.RepositoryCollectionCapability.*;
 import static org.molgenis.data.meta.AttributeType.*;
-import static org.molgenis.data.meta.MetaUtils.getEntityTypeFetch;
 import static org.molgenis.data.meta.model.EntityTypeMetadata.*;
 import static org.molgenis.data.postgresql.PostgreSqlQueryGenerator.*;
 import static org.molgenis.data.postgresql.PostgreSqlQueryUtils.*;
 import static org.molgenis.data.postgresql.PostgreSqlRepository.BATCH_SIZE;
 import static org.molgenis.data.postgresql.PostgreSqlRepository.createJunctionTableRowData;
-import static org.molgenis.data.support.EntityTypeUtils.*;
+import static org.molgenis.data.util.EntityTypeUtils.*;
+import static org.molgenis.data.util.MetaUtils.getEntityTypeFetch;
 import static org.molgenis.util.stream.MapCollectors.toLinkedMap;
 import static org.springframework.jdbc.support.JdbcUtils.closeConnection;
