+++ molgenis-data-postgresql/src/main/java/org/molgenis/data/postgresql/identifier/EntityTypeRegistryImpl.java
@@ -22,8 +22,8 @@ import static com.google.common.collect.Streams.stream;
 import static java.util.stream.Collectors.toMap;
 import static java.util.stream.Collectors.toSet;
 import static org.molgenis.data.postgresql.PostgreSqlNameGenerator.getJunctionTableName;
-import static org.molgenis.data.support.EntityTypeUtils.isMultipleReferenceType;
 import static org.molgenis.data.transaction.TransactionManager.TRANSACTION_ID_RESOURCE_NAME;
+import static org.molgenis.data.util.EntityTypeUtils.isMultipleReferenceType;
 
 /**
  * Tracks PostgreSQL table names.
