+++ molgenis-data-rest/src/main/java/org/molgenis/data/rest/v2/AttributeResponseV2.java
@@ -7,7 +7,7 @@ import org.molgenis.data.meta.AttributeType;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.rest.util.Href;
-import org.molgenis.data.support.EntityTypeUtils;
+import org.molgenis.data.util.EntityTypeUtils;
 import org.molgenis.security.core.UserPermissionEvaluator;
 
 import java.util.List;
