+++ molgenis-data-rest/src/main/java/org/molgenis/data/rest/v2/EntityTypeResponseV2.java
@@ -9,7 +9,7 @@ import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.rest.util.Href;
 import org.molgenis.data.security.EntityTypeIdentity;
 import org.molgenis.data.security.EntityTypePermission;
-import org.molgenis.data.support.EntityTypeUtils;
+import org.molgenis.data.util.EntityTypeUtils;
 import org.molgenis.security.core.Permission;
 import org.molgenis.security.core.UserPermissionEvaluator;
