+++ molgenis-data/src/main/java/org/molgenis/data/populate/DefaultValuePopulator.java
@@ -6,7 +6,7 @@ import org.molgenis.data.meta.AttributeType;
 import org.molgenis.data.meta.IllegalAttributeTypeException;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
-import org.molgenis.data.support.AttributeUtils;
+import org.molgenis.data.util.AttributeUtils;
 import org.molgenis.util.ListEscapeUtils;
 import org.molgenis.util.UnexpectedEnumException;
 import org.springframework.stereotype.Component;
