+++ molgenis-data/src/main/java/org/molgenis/data/support/DynamicEntity.java
@@ -6,6 +6,7 @@ import org.molgenis.data.UnknownAttributeException;
 import org.molgenis.data.meta.AttributeType;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
+import org.molgenis.data.util.EntityTypeUtils;
 import org.molgenis.util.UnexpectedEnumException;
 
 import java.time.Instant;
