+++ molgenis-data/src/main/java/org/molgenis/data/support/LazyEntity.java
@@ -5,6 +5,7 @@ import org.molgenis.data.Entity;
 import org.molgenis.data.UnknownEntityException;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
+import org.molgenis.data.util.EntityTypeUtils;
 
 import java.time.Instant;
 import java.time.LocalDate;
