+++ molgenis-metadata-manager/src/test/java/org/molgenis/metadata/manager/mapper/EntityTypeMapperTest.java
@@ -18,7 +18,7 @@ import java.util.Map;
 import static com.google.common.collect.ImmutableList.of;
 import static java.util.Collections.singletonMap;
 import static org.mockito.Mockito.*;
-import static org.molgenis.data.support.AttributeUtils.getI18nAttributeName;
+import static org.molgenis.data.util.AttributeUtils.getI18nAttributeName;
 import static org.testng.Assert.assertEquals;
 
 public class EntityTypeMapperTest
