+++ molgenis-semantic-mapper/src/main/java/org/molgenis/semanticmapper/controller/MappingServiceController.java
@@ -15,8 +15,8 @@ import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.meta.model.Package;
 import org.molgenis.data.semantic.Relation;
 import org.molgenis.data.support.AggregateQueryImpl;
-import org.molgenis.data.support.EntityTypeUtils;
 import org.molgenis.data.support.QueryImpl;
+import org.molgenis.data.util.EntityTypeUtils;
 import org.molgenis.dataexplorer.controller.DataExplorerController;
 import org.molgenis.jobs.JobExecutor;
 import org.molgenis.ontology.core.model.OntologyTerm;
@@ -59,8 +59,8 @@ import static java.text.MessageFormat.format;
 import static java.util.stream.Collectors.toList;
 import static org.apache.commons.lang3.StringUtils.trim;
 import static org.molgenis.data.meta.AttributeType.COMPOUND;
-import static org.molgenis.data.meta.MetaUtils.isSystemPackage;
 import static org.molgenis.data.rest.util.Href.concatEntityHref;
+import static org.molgenis.data.util.MetaUtils.isSystemPackage;
 import static org.molgenis.data.validation.meta.NameValidator.validateEntityName;
 import static org.molgenis.security.core.utils.SecurityUtils.currentUserIsSu;
 import static org.molgenis.security.core.utils.SecurityUtils.getCurrentUsername;
